extracted stats methods into stats helper module
