+++ app/models/dog.rb
@@ -1,4 +1,5 @@
 class Dog < ActiveRecord::Base
+  include StatsHelper
 
   has_attached_file :avatar,
   :styles => { :thumb => '60x60#', :medium => '200x200#', :large => '300x300#' }, :default_style => :large,
@@ -25,18 +26,5 @@ class Dog < ActiveRecord::Base
       return true
     end
   end
-
-  def recent_walks
-    self.walks.order(:created_at).limit(5)
-  end
-
-  def distance_traveled
-     distance = 0
-     self.walks.each do |walk|
-       distance += walk.distance_traveled
-     end
-     distance.round(1)
-  end
-
 
 end
