+++ app/models/user.rb
@@ -1,4 +1,5 @@
 class User < ActiveRecord::Base
+  include StatsHelper
 
   has_attached_file :avatar,
   :styles => { :thumb => '60x60#', :medium => '200x200#', :large => '300x300#' }, :default_style => :large,
@@ -16,31 +17,6 @@ class User < ActiveRecord::Base
 
   has_secure_password
 
-  def recent_walks
-    self.walks.order(:created_at).limit(5)
-  end
-
-  def distance_traveled
-     distance = 0
-     self.walks.each do |walk|
-       distance += walk.distance_traveled
-     end
-     distance.round(1)
-  end
-
-
-  def distance_score
-    (distance_traveled * 250).to_i
-  end
-
-
-  def area
-    area = 0
-    self.walks.each do |walk|
-      area += walk.area.to_f
-    end
-    area
-  end
 
 
 end
